lsnes rr1-Δ17ε3

+++ src/core/movie.cpp
@@ -574,7 +574,7 @@ void movie::adjust_frame_count(int64_t adjust)
 				current_frame_first_subframe += count_changes(current_frame_first_subframe);
 	}
 	//Nobody is this stupid, right?
-	current_frame_first_subframe = min(current_frame_first_subframe, movie_data.size());
+	current_frame_first_subframe = min(current_frame_first_subframe, static_cast<uint64_t>(movie_data.size()));
 }
 
 unsigned extended_mode = 0;
